made docker installation optional

+++ scripts/docker_install.sh
@@ -20,21 +20,23 @@ if [[ $1 == "f" ]]; then
         /home/ridonk/media/music \
         /home/ridonk/media/tvseries
 fi
-sudo apt-get remove docker docker-engine docker.io containerd runc
-sudo apt-get update && sudo apt-get upgrade
-sudo apt-get install \
-    apt-transport-https \
-    ca-certificates \
-    curl \
-    software-properties-common
- curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo add-apt-repository \
-   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-   $(lsb_release -cs) \
-   stable"
-sudo apt-get update
-sudo apt-get install docker-ce
-sudo groupadd docker
-sudo usermod -aG docker $USER
-sudo systemctl enable docker
-logout
+if [[ $1 == "r" ]]; then
+    sudo apt-get remove docker docker-engine docker.io containerd runc
+    sudo apt-get update && sudo apt-get upgrade
+    sudo apt-get install \
+        apt-transport-https \
+        ca-certificates \
+        curl \
+        software-properties-common
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+    sudo add-apt-repository \
+    "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+    $(lsb_release -cs) \
+    stable"
+    sudo apt-get update
+    sudo apt-get install docker-ce
+    sudo groupadd docker
+    sudo usermod -aG docker $USER
+    sudo systemctl enable docker
+    logout
+fi
